+++ backend/canisters/user_index/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Ensure canisters within `include` are always relayed on to local indexes ([#8083](https://github.com/open-chat-labs/open-chat/pull/8083))
+- Add missing decoding for community read permissions ([#8141](https://github.com/open-chat-labs/open-chat/pull/8141))
 
 ## [[2.0.1776](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1776-user_index)] - 2025-06-02